pyproject.toml
[project]
name = "tavern"
version = "0.1.0"
description = "CLI interactive novel game — fantasy tavern exploration"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0,<3.0",
    "rich>=13.0",
    "networkx>=3.0",
    "openai>=1.0",
    "pyyaml>=6.0",
    "tenacity>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[project.scripts]
tavern = "tavern.__main__:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
